requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.2.1
chardet==3.0.2
discord.py==0.16.7
multidict==2.1.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
websockets==3.3
-e git://github.com/MemorableUsername/butter#egg=butter3
